Kleinkram

+++ update_classes.bat
@@ -1,5 +1,7 @@
 @echo off
-cd %~dp0
+cd /d %~dp0
+net session >nul 2>&1
+if not %errorlevel%==0 goto :abort
 if exist temp rmdir /s /q temp> nul
 mkdir temp\tex\latex\tudscr\logo
 xcopy source temp\
@@ -14,3 +16,14 @@ xcopy tex "%texmfpath%\tex\" /s /y
 cd %~dp0
 if exist temp rmdir /s /q temp> nul
 texhash
+exit /b 0
+:abort
+echo.
+echo  =====================================================================
+echo.
+echo   Installation muss als Admin erfolgen
+echo.
+echo  =====================================================================
+echo.
+pause
+exit /b 0
